+++ test/datetime_extension_test.dart
@@ -62,4 +62,27 @@ void main() {
     expect(winter.season, Season.winter);
     expect(winterEnd.season, Season.winter);
   });
+
+  test('timeAgo', () {
+    final DateTime now = DateTime.now();
+    final DateTime dt1 = now.add(const Duration(days: -(365 * 3)));
+    final DateTime dt2 = now.add(const Duration(days: -366));
+    final DateTime dt3 = now.add(const Duration(days: -(365 ~/ 2)));
+    final DateTime dt4 = now.add(const Duration(days: -14));
+    final DateTime dt5 = now.add(const Duration(days: -8));
+    final DateTime dt6 = now.add(const Duration(hours: -5));
+    final DateTime dt7 = now.add(const Duration(hours: -1));
+    final DateTime dt8 = now.add(const Duration(minutes: -5));
+    final DateTime dt9 = now.add(const Duration(seconds: -5));
+
+    expect(dt1.timeAgo, '3 years ago');
+    expect(dt2.timeAgo, '1 year ago');
+    expect(dt3.timeAgo, '6 months ago');
+    expect(dt4.timeAgo, '2 weeks ago');
+    expect(dt5.timeAgo, '1 week ago');
+    expect(dt6.timeAgo, '5 hours ago');
+    expect(dt7.timeAgo, '1 hour ago');
+    expect(dt8.timeAgo, '5 minutes ago');
+    expect(dt9.timeAgo, 'just now');
+  });
 }